bugfix(gui): Fix application hang from scrolling a map list with arrow keys (#2281)

+++ Generals/Code/GameEngine/Include/GameClient/Gadget.h
@@ -371,11 +371,11 @@ typedef struct _ListboxData
 	Int					selectPos;				// Position of current selected entry (for SINGLE select)
 	Int					*selections;			// Pointer to array of selections (for MULTI select)
 
-	Short				displayHeight;		// Height in pixels of listbox display region
+	Int					displayHeight;		// Height in pixels of listbox display region
 																// this is computed based on the existence
 																// of a title or not.
 	UnsignedInt doubleClickTime;	//
-	Short				displayPos;				// Position of current display entry in pixels
+	Int					displayPos;				// Position of current display entry in pixels
 
 } ListboxData;
